Assignment 3/app/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.1.5
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==2.0.3
beautifulsoup4==4.12.3
python-magic==0.4.27
defusedxml==0.7.1
lxml==5.3.0